Renamed `Selector` module to `Polling`
