+++ examples/1/tcp/main.cpp
@@ -6,7 +6,7 @@
 #include <string>
 #include <thread>
 
-#include <Sockets/selector.hpp>
+#include <Sockets/polling.hpp>
 #include <Sockets/socket.hpp>
 
 std::string process(std::shared_ptr<Sockets::Socket> p) {
