+++ socket/CMakeLists.txt
@@ -2,4 +2,4 @@ cmake_minimum_required(VERSION 3.16)
 
 add_subdirectory(Socket)
 add_subdirectory(ThreadPool)
-add_subdirectory(Selector)
+add_subdirectory(Polling)
